move InsertColumnValuesTest.java
